Extract assistant context assembler
