+++ .codex-supervisor/issues/455/issue-journal.md
@@ -1,0 +1,33 @@
+# Issue #455: implementation: extract assistant-context and advisory assembly from the monolithic service
+
+## Supervisor Snapshot
+- Issue URL: https://github.com/TommyKammy/AegisOps/issues/455
+- Branch: codex/issue-455
+- Workspace: .
+- Journal: .codex-supervisor/issues/455/issue-journal.md
+- Current phase: reproducing
+- Attempt count: 1 (implementation=1, repair=0)
+- Last head SHA: 79d4dd5af3a76b89542f34e4227e865632831417
+- Blocked reason: none
+- Last failure signature: none
+- Repeated failure signature count: 0
+- Updated at: 2026-04-14T00:18:21.195Z
+
+## Latest Codex Summary
+- None yet.
+
+## Active Failure Context
+- None recorded.
+
+## Codex Working Notes
+### Current Handoff
+- Hypothesis: The issue can be satisfied by extracting assistant-context and advisory assembly into a dedicated internal collaborator while keeping `AegisOpsControlPlaneService` as a public facade and preserving the existing Phase 19 fail-closed enforcement.
+- What changed: Added a focused delegation test, introduced `control-plane/aegisops_control_plane/assistant_context.py` with `AssistantContextAssembler`, wired the service to delegate `inspect_assistant_context`, `inspect_advisory_output`, and `render_recommendation_draft`, and moved the assistant/advisory assembly helper block out of `service.py`.
+- Current blocker: none
+- Next exact step: Stage the extracted collaborator, service facade change, test update, and journal update, then create a checkpoint commit on `codex/issue-455`.
+- Verification gap: none for the required local verification commands; only manual reviewer inspection remains for service body dominance.
+- Files touched: `control-plane/aegisops_control_plane/assistant_context.py`, `control-plane/aegisops_control_plane/service.py`, `control-plane/tests/test_service_persistence.py`, `.codex-supervisor/issues/455/issue-journal.md`
+- Rollback concern: Low-to-moderate; the extraction is behavior-preserving but the collaborator still depends on many service-private helpers, so regressions would likely come from wiring rather than policy logic.
+- Last focused command: `python3 -m unittest control-plane.tests.test_service_persistence control-plane.tests.test_phase19_operator_workflow_validation`
+### Scratchpad
+- Keep this section short. The supervisor may compact older notes automatically.
